Split up multibody_tree.h.
