+++ multibody/tree/acceleration_kinematics_cache.h
@@ -32,6 +32,7 @@ namespace internal {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 template <typename T>
