+++ multibody/tree/articulated_body_inertia.h
@@ -93,6 +93,7 @@ namespace multibody {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 template<typename T>
