+++ multibody/tree/fixed_offset_frame.h
@@ -6,7 +6,6 @@
 #include "drake/common/default_scalars.h"
 #include "drake/common/eigen_types.h"
 #include "drake/multibody/tree/frame.h"
-#include "drake/multibody/tree/multibody_tree_forward_decl.h"
 #include "drake/multibody/tree/multibody_tree_topology.h"
 
 namespace drake {
