+++ multibody/tree/linear_spring_damper.h
@@ -44,6 +44,7 @@ template <typename T> class Body;
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 /// No other values for T are currently supported.
