+++ multibody/tree/multibody_tree_context.h
@@ -36,6 +36,7 @@ namespace internal {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 template <typename T>
@@ -82,18 +83,12 @@ class MultibodyTreeContext final : public systems::LeafContext<T> {
     return topology_.num_velocities();
   }
 
-  /// Returns `true` if the state is discrete and `false` if the state is
-  /// continuous.
-  bool is_state_discrete() const {
-    return is_state_discrete_;
-  }
-
   /// Returns a const reference to the state vector stored in `this` context as
   /// an `Eigen::VectorBlock<const VectorX<T>>`.
   Eigen::VectorBlock<const VectorX<T>> get_state_vector() const {
     DRAKE_ASSERT(this->get_num_discrete_state_groups() <= 1);
     const systems::BasicVector<T>& state_vector =
-        is_state_discrete() ? this->get_discrete_state(0) :
+        is_state_discrete_ ? this->get_discrete_state(0) :
         dynamic_cast<const systems::BasicVector<T>&>(
             this->get_continuous_state().get_vector());
     return state_vector.get_value();
@@ -113,10 +108,10 @@ class MultibodyTreeContext final : public systems::LeafContext<T> {
   Eigen::VectorBlock<VectorX<T>> get_mutable_state_vector(
       systems::State<T>* state) const {
     DRAKE_ASSERT(this->get_num_discrete_state_groups() <= 1);
-    DRAKE_ASSERT(is_state_discrete() ==
+    DRAKE_ASSERT(is_state_discrete_ ==
                  (this->get_num_discrete_state_groups() == 1));
     systems::BasicVector<T>& state_vector =
-        is_state_discrete() ? state->get_mutable_discrete_state(0) :
+        is_state_discrete_ ? state->get_mutable_discrete_state(0) :
         dynamic_cast<systems::BasicVector<T>&>(
             state->get_mutable_continuous_state().get_mutable_vector());
     return state_vector.get_mutable_value();
