+++ multibody/tree/quaternion_floating_mobilizer.h
@@ -33,6 +33,7 @@ namespace internal {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 /// No other values for T are currently supported.
