+++ multibody/tree/test/free_rotating_body_plant.h
@@ -25,6 +25,7 @@ namespace test {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 /// No other values for T are currently supported.
@@ -106,3 +107,6 @@ class FreeRotatingBodyPlant final : public internal::MultibodyTreeSystem<T> {
 }  // namespace test
 }  // namespace multibody
 }  // namespace drake
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
+    class ::drake::multibody::test::FreeRotatingBodyPlant)
