+++ multibody/tree/velocity_kinematics_cache.h
@@ -36,6 +36,7 @@ namespace internal {
 ///
 /// - double
 /// - AutoDiffXd
+/// - symbolic::Expression
 ///
 /// They are already available to link against in the containing library.
 template <typename T>
